Add ranking and voting modules, integrating model metadata from configuration. Refactor UI components for voting and ranking display, enhancing user interaction and data management.
