+++ app.py
@@ -19,6 +19,26 @@ from elo_calculation import (
     create_ranking_dataframe,
 )
 
+# Import ranking functionality
+from ranking import (
+    load_ranking_data,
+    update_ranking_display,
+    force_update_ranking_display,
+    create_ranking_tab,
+    setup_ranking_handlers,
+)
+
+# Import voting functionality
+from voting import (
+    handle_vote,
+    save_vote_to_hf,
+    serialize_interactions,
+    create_vote_ui,
+    should_show_vote_buttons,
+    get_vote_ui_updates,
+    setup_vote_handlers,
+)
+
 # Import completion utilities
 from completion import make_config, registered_api_completion
 from sandbox.prompts import GENERAL_SANDBOX_INSTRUCTION
@@ -102,9 +122,6 @@ available_models = list(api_config.keys()) if api_config else []
 HF_DATASET_NAME = os.getenv("HF_DATASET_NAME")
 HF_TOKEN = os.getenv("HF_TOKEN")
 
-# Global ranking data cache
-ranking_data = None
-ranking_last_updated = None
 
 def get_random_models():
     """Get two random models from available models"""
@@ -560,14 +577,7 @@ def retry_last_message(state0, state1, model_a, model_b):
     new_state0, new_state1 = result[0], result[1]
 
     # Check if both models have output and are not generating to show vote buttons
-    show_vote_buttons = (
-        new_state0
-        and new_state0.get("has_output", False)
-        and not new_state0.get("generating", False)
-        and new_state1
-        and new_state1.get("has_output", False)
-        and not new_state1.get("generating", False)
-    )
+    show_vote_buttons = should_show_vote_buttons(new_state0, new_state1)
 
     # Return all the original outputs plus the updated state for run buttons
     return (
@@ -814,90 +824,6 @@ def send_to_right_only(state0, state1, text, temperature, max_tokens, model_a, m
     )
 
 
-def handle_vote(state0, state1, vote_type):
-    """Handle vote submission"""
-    if (
-        not state0
-        or not state1
-        or not state0.get("has_output")
-        or not state1.get("has_output")
-    ):
-        return (
-            "No output to vote on!",
-            gr.update(),
-            "**Last Updated:** No data available",
-        )
-
-    # Get all user messages and the last responses
-    user_messages = []
-    response_a = ""
-    response_b = ""
-
-    # Collect all user messages from the conversation
-    for msg in state0["messages"]:
-        if msg["role"] == "user":
-            user_messages.append(msg["content"])
-
-    for msg in reversed(state0["messages"]):
-        if msg["role"] == "assistant":
-            response_a = msg["content"]
-            break
-
-    for msg in reversed(state1["messages"]):
-        if msg["role"] == "assistant":
-            response_b = msg["content"]
-            break
-
-    # Get interactions and full conversation history for remote dataset saving
-    interactions_a = state0.get("interactions", [])
-    interactions_b = state1.get("interactions", [])
-    
-    # Get full conversation history for both models
-    conversation_a = state0.get("messages", [])
-    conversation_b = state1.get("messages", [])
-    
-    # Save vote with full conversation history to remote dataset in background (async)
-    import threading
-    def save_vote_background():
-        try:
-            success, message = save_vote_to_hf(
-                state0["model_name"],
-                state1["model_name"],
-                user_messages[0],
-                response_a,
-                response_b,
-                vote_type,
-                interactions_a,
-                interactions_b,
-                conversation_a,
-                conversation_b,
-            )
-
-        except Exception as e:
-            print(f"Error saving vote: {str(e)}")
-            pass
-    
-    print("Saving vote in background...")
-    # Start background upload thread
-    upload_thread = threading.Thread(target=save_vote_background)
-    upload_thread.daemon = True
-    upload_thread.start()
-    
-    # Return immediately without waiting for upload
-    success = True  # Assume success for immediate UI response
-    message = "Vote recorded! Uploading data in background..."
-
-    if success:
-        # Return immediately without waiting for ranking refresh
-        return (
-            message + " Clearing conversation...",
-            gr.update(),  # Keep existing ranking table
-            "**Last Updated:** Processing in background...",
-        )
-    else:
-        return message, gr.update(), "**Last Updated:** Error occurred"
-
-
 def run_sandbox_code(sandbox_state: dict, code: str, install_command: str) -> tuple[str, str, str]:
     """Run code in the appropriate sandbox environment"""
     if not code.strip():
@@ -1000,213 +926,28 @@ async def run_sandbox_code_async(sandbox_state: dict, code: str, install_command
 async def run_sandboxes_parallel(sandbox_state0, code0, install_command0, sandbox_state1, code1, install_command1):
     """Run both sandbox executions in parallel with error handling"""
     loop = asyncio.get_event_loop()
-    
+
     # Create tasks for both sandbox executions
     task0 = loop.run_in_executor(None, run_sandbox_code, sandbox_state0, code0, install_command0)
     task1 = loop.run_in_executor(None, run_sandbox_code, sandbox_state1, code1, install_command1)
-    
+
     # Wait for both to complete with error handling
     try:
         result0, result1 = await asyncio.gather(task0, task1, return_exceptions=True)
-        
+
         # Handle exceptions
         if isinstance(result0, Exception):
             result0 = ("", "", f"Sandbox execution error: {str(result0)}")
-        
+
         if isinstance(result1, Exception):
             result1 = ("", "", f"Sandbox execution error: {str(result1)}")
-            
+
     except Exception as e:
         # Fallback to sequential processing
         result0 = run_sandbox_code(sandbox_state0, code0, install_command0)
         result1 = run_sandbox_code(sandbox_state1, code1, install_command1)
-    
+
     return result0, result1
-
-
-def serialize_interactions(interactions):
-    """Convert datetime objects in interactions to ISO format strings"""
-    if not interactions:
-        return interactions
-    
-    serialized = []
-    for interaction in interactions:
-        # Handle case where interaction might be a list instead of a dict
-        if isinstance(interaction, list):
-            # If it's a list, recursively serialize each item
-            serialized.append(serialize_interactions(interaction))
-        elif isinstance(interaction, dict):
-            # If it's a dict, serialize it normally
-            serialized_interaction = {}
-            for key, value in interaction.items():
-                if isinstance(value, datetime.datetime):
-                    serialized_interaction[key] = value.isoformat()
-                else:
-                    serialized_interaction[key] = value
-            serialized.append(serialized_interaction)
-        else:
-            # If it's neither list nor dict, just add it as is
-            serialized.append(interaction)
-    return serialized
-
-
-def save_vote_to_hf(
-    model_a, model_b, prompt, response_a, response_b, vote_result, interactions_a=None, interactions_b=None, conversation_a=None, conversation_b=None, hf_token=None
-):
-    """Save vote result to HuggingFace dataset with full conversation history"""
-    try:
-        # Use global token if not provided
-        token = hf_token or HF_TOKEN
-        if not token:
-            return False, "HuggingFace token not found in environment (HF_TOKEN)"
-
-        if not HF_DATASET_NAME:
-            return False, "HuggingFace dataset name not found in environment (HF_DATASET_NAME)"
-
-        # Serialize conversations for JSON compatibility
-        serialized_conversation_a = serialize_interactions(conversation_a or [])
-        serialized_conversation_b = serialize_interactions(conversation_b or [])
-
-        # Organize interactions by turns - each turn contains a list of interactions
-        def organize_interactions_by_turns(interactions, conversation):
-            """Organize interactions by conversation turns"""
-            if not interactions:
-                return []
-
-            # For now, put all interactions in a single turn
-            # This can be enhanced later to properly group by conversation turns
-            # when we have more context about how interactions are timestamped
-            return interactions if interactions else []
-
-        # Organize interactions by turns for both models
-        action_a = organize_interactions_by_turns(interactions_a or [], conversation_a or [])
-        action_b = organize_interactions_by_turns(interactions_b or [], conversation_b or [])
-
-        # Serialize actions for JSON compatibility
-        serialized_action_a = serialize_interactions(action_a)
-        serialized_action_b = serialize_interactions(action_b)
-
-        # Create vote data with full conversation history and actions organized by turns
-        # Each conversation is a list of messages in format: [{"role": "user"/"assistant", "content": "...", "action": [...]}, ...]
-        # Actions are organized as list of lists: [[turn1_interactions], [turn2_interactions], ...]
-        vote_data = {
-            "timestamp": datetime.datetime.now().isoformat(),
-            "model_a": model_a,
-            "model_b": model_b,
-            "initial_prompt": prompt,  # Convert list to single string
-            "action_a": serialized_action_a,  # Actions organized by turns for model A
-            "action_b": serialized_action_b,  # Actions organized by turns for model B
-            "conversation_a": serialized_conversation_a,  # Full conversation history for model A
-            "conversation_b": serialized_conversation_b,  # Full conversation history for model B
-            "vote": vote_result,  # "left", "right", "tie", "both_bad"
-        }
-
-        # Try to load existing dataset or create new one
-        try:
-            dataset = load_dataset(HF_DATASET_NAME, split="train", token=token)
-            # Convert to pandas DataFrame - handle both Dataset and DatasetDict
-            if hasattr(dataset, "to_pandas"):
-                df = dataset.to_pandas()
-            else:
-                df = pd.DataFrame(dataset)
-            # Add new vote
-            new_df = pd.concat([df, pd.DataFrame([vote_data])], ignore_index=True)
-        except Exception as load_error:
-            # Create new dataset if it doesn't exist
-            new_df = pd.DataFrame([vote_data])
-
-        # Convert back to dataset and push
-        new_dataset = Dataset.from_pandas(new_df)
-        try:
-            new_dataset.push_to_hub(HF_DATASET_NAME, token=token)
-            return True, "Vote saved successfully!"
-        except Exception as upload_error:
-            return False, f"Error uploading to HuggingFace: {str(upload_error)}"
-    except Exception as e:
-        return False, f"Error saving vote: {str(e)}"
-
-
-def load_ranking_data(hf_token=None, force_reload=False):
-    """Load and calculate ranking data from HuggingFace dataset"""
-    global ranking_data, ranking_last_updated
-
-    try:
-        # Use global token if not provided
-        token = hf_token or HF_TOKEN
-        if not token:
-            return pd.DataFrame()
-
-        # Load dataset - force download if requested
-        if force_reload:
-            # Force download from remote, ignore cache
-            dataset = load_dataset(
-                HF_DATASET_NAME,
-                split="train",
-                token=token,
-                download_mode="force_redownload",
-            )
-        else:
-            dataset = load_dataset(HF_DATASET_NAME, split="train", token=token)
-        # Convert to pandas DataFrame - handle both Dataset and DatasetDict
-        if hasattr(dataset, "to_pandas"):
-            df = dataset.to_pandas()
-        else:
-            df = pd.DataFrame(dataset)
-
-        if df.empty:
-            return pd.DataFrame()
-
-        # Convert vote format for Elo calculation and count votes
-        battle_data = []
-        vote_counts = defaultdict(int)
-
-        for _, row in df.iterrows():
-            model_a = row["model_a"]
-            model_b = row["model_b"]
-            vote = row["vote"]
-
-            # Convert vote to winner format for Elo
-            if vote == "left":  # Model A wins
-                winner = "model_a"
-            elif vote == "right":  # Model B wins
-                winner = "model_b"
-            elif vote == "tie":
-                winner = "tie"
-            elif vote == "both_bad":
-                winner = "tie (bothbad)"
-            else:
-                continue  # Skip invalid votes
-
-            battle_data.append(
-                {"model_a": model_a, "model_b": model_b, "winner": winner}
-            )
-
-            # Count votes for each model
-            vote_counts[model_a] += 1
-            vote_counts[model_b] += 1
-
-        # Create DataFrame for Elo calculation
-        battles_df = pd.DataFrame(battle_data)
-
-        if battles_df.empty:
-            return pd.DataFrame()
-
-        # Calculate Elo ratings using Bradley-Terry Model with confidence intervals
-        elo_ratings, confidence_intervals = calculate_elo_with_confidence_intervals(
-            battles_df, vote_counts
-        )
-
-        # Create ranking DataFrame
-        ranking_df = create_ranking_dataframe(
-            elo_ratings, confidence_intervals, vote_counts
-        )
-
-        ranking_data = ranking_df
-        ranking_last_updated = datetime.datetime.now()
-
-        return ranking_df
-    except Exception as e:
-        return pd.DataFrame()
 
 
 def instantiate_send_button():
@@ -1368,25 +1109,15 @@ def build_ui():
                                         interactive=False,
                                     )
 
-                # Vote buttons section - only visible after output
-                with gr.Row(visible=False) as vote_section:
-                    gr.Markdown("### 🗳️ Which response is better?")
-                with gr.Row(visible=False) as vote_buttons_row:
-                    vote_left_btn = gr.Button(
-                        "👍 A is Better", variant="primary", size="lg"
-                    )
-                    vote_tie_btn = gr.Button(
-                        "🤝 It's a Tie", variant="secondary", size="lg"
-                    )
-                    vote_both_bad_btn = gr.Button(
-                        "👎 Both are Bad", variant="secondary", size="lg"
-                    )
-                    vote_right_btn = gr.Button(
-                        "👍 B is Better", variant="primary", size="lg"
-                    )
-
-                # Vote status message
-                vote_status = gr.Markdown("", visible=False)
+                # Vote UI components
+                vote_components = create_vote_ui()
+                vote_section = vote_components["vote_section"]
+                vote_buttons_row = vote_components["vote_buttons_row"]
+                vote_left_btn = vote_components["vote_left_btn"]
+                vote_right_btn = vote_components["vote_right_btn"]
+                vote_tie_btn = vote_components["vote_tie_btn"]
+                vote_both_bad_btn = vote_components["vote_both_bad_btn"]
+                vote_status = vote_components["vote_status"]
 
                 # Main chat interface - Collapsible and hidden by default
                 with gr.Accordion("💬 Chat Interface", open=False):
@@ -1575,34 +1306,7 @@ def build_ui():
                     inputs=[text_input],
                 )
             # Ranking Tab
-            with gr.Tab("📊 Ranking", id="ranking"):
-                gr.Markdown("## 🏆 Model Leaderboard")
-                gr.Markdown("*Rankings auto-refresh every 10 minutes*")
-
-                ranking_table = gr.Dataframe(
-                    headers=[
-                        "Rank",
-                        "Model",
-                        "Score",
-                        "95% CI (±)",
-                        "Votes",
-                    ],
-                    datatype=[
-                        "number",
-                        "str",
-                        "number",
-                        "str",
-                        "number",
-                    ],
-                    label="Model Rankings",
-                    interactive=False,
-                    wrap=True,
-                )
-
-                ranking_last_update = gr.Markdown("**Last Updated:** Not loaded yet")
-
-                # Timer for auto-refresh every 10 minutes
-                ranking_timer = gr.Timer(value=600.0, active=True)
+            ranking_table, ranking_last_update, ranking_timer = create_ranking_tab()
 
         # Event handlers
         # Create state variables for the run buttons
@@ -1662,14 +1366,7 @@ def build_ui():
             new_state0, new_state1 = result[0], result[1]
 
             # Check if both models have output and are not generating to show vote buttons
-            show_vote_buttons = (
-                new_state0
-                and new_state0.get("has_output", False)
-                and not new_state0.get("generating", False)
-                and new_state1
-                and new_state1.get("has_output", False)
-                and not new_state1.get("generating", False)
-            )
+            show_vote_buttons = should_show_vote_buttons(new_state0, new_state1)
 
             # Return all the original outputs plus the updated state for run buttons
             # Make sure all outputs are properly formatted for their expected types
@@ -2042,7 +1739,7 @@ def build_ui():
             ],
         )
 
-        # Vote button handlers
+        # Setup vote handlers
         def process_vote(state0, state1, vote_type, current_text):
             # Save the vote and get updates
             message, ranking_update, last_update = handle_vote(
@@ -2129,45 +1826,8 @@ def build_ui():
                 ],
             )
 
-        # Ranking handlers
-        def update_ranking_display():
-            df = load_ranking_data()
-            if df.empty:
-                return gr.update(value=df), "**Last Updated:** No data available"
-
-            last_update = (
-                ranking_last_updated.strftime("%Y-%m-%d %H:%M:%S")
-                if ranking_last_updated
-                else "Unknown"
-            )
-            return gr.update(value=df), f"**Last Updated:** {last_update}"
-
-        def force_update_ranking_display():
-            """Force update ranking data from HuggingFace (for timer)"""
-            df = load_ranking_data(force_reload=True)
-            if df.empty:
-                return gr.update(value=df), "**Last Updated:** No data available"
-
-            last_update = (
-                ranking_last_updated.strftime("%Y-%m-%d %H:%M:%S")
-                if ranking_last_updated
-                else "Unknown"
-            )
-            return gr.update(value=df), f"**Last Updated:** {last_update}"
-
-        # Timer tick handler for auto-refresh with force reload
-        ranking_timer.tick(
-            fn=force_update_ranking_display,
-            inputs=[],
-            outputs=[ranking_table, ranking_last_update],
-        )
-
-        # Auto-load ranking on startup
-        demo.load(
-            fn=update_ranking_display,
-            inputs=[],
-            outputs=[ranking_table, ranking_last_update],
-        )
+        # Setup ranking handlers
+        setup_ranking_handlers(demo, ranking_table, ranking_last_update, ranking_timer)
 
     return demo
